Card discount type changed

+++ src/main/java/com/kkotto/Clevertec/service/model/entity/Card.java
@@ -14,7 +14,7 @@ public class Card {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String ownerName;
-    private Integer discount;
+    private double discount;
     private Integer cardLastDigits;
     private LocalDateTime createAt;
